Refactor shade persistence into dedicated runtime
